--- web/src/pages/home/style.css
@@ -1,22 +1,0 @@
-/* .appHeader {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding-bottom: 1rem;
-}
-
-.appHeader > p {
-  margin: 0;
-  font-size: 1.5rem;
-}
-
-.emptyState {
-  padding: 5rem;
-}
-
-.loader {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  padding: 5rem;
-} */
