--- web/src/service/app/index.ts
@@ -1,1 +1,0 @@
-export { fetchActions, saveBatch } from "./action_group";
