+++ web/src/stores/route.store.js
@@ -1,0 +1,8 @@
+import { create } from "zustand";
+
+export const routeStore = create((set) => ({
+  appActiveMenu: "dashboard",
+  setAppActiveMenu: (menu) => set({ appActiveMenu: menu })
+}));
+
+
