--- web/src/stores/route.store.ts
@@ -1,8 +1,0 @@
-import { create } from "zustand";
-
-export const orcaStore = create((set) => ({
-  appActiveMenu: "dashboard",
-  setAppActiveMenu: (menu: string) => set({ appActiveMenu: menu })
-}));
-
-
